+++ src/platform/Darwin/BleApplicationDelegateImpl.mm
@@ -1,6 +1,6 @@
 /*
  *
- *    Copyright (c) 2020 Project CHIP Authors
+ *    Copyright (c) 2020-2025 Project CHIP Authors
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -25,7 +25,7 @@
 #endif
 
 #include <ble/Ble.h>
-#include <platform/Darwin/BleApplicationDelegate.h>
+#include <platform/Darwin/BleApplicationDelegateImpl.h>
 
 using namespace ::chip;
 using namespace ::chip::Ble;
