+++ src/platform/Darwin/BlePlatformDelegateImpl.h
